ESC-465 move test/utils into test project hierarchy
